// types/tokenfactory/arb-vault/eris_arb_vault_instantiate.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0
 *
 * The greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)
 */
export type Decimal = string;
export type LsdTypeFor_String =
  | {
      eris: {
        addr: string;
        denom: string;
      };
    }
  | {
      backbone: {
        addr: string;
        denom: string;
      };
    };
export type UtilizationMethod = {
  steps: [Decimal, Decimal][];
};

export interface InstantiateMsg {
  /**
   * Denom of the arb[TOKEN] e.g. arbWHALE
   */
  denom: string;
  fee_config: FeeConfigFor_String;
  lsds: LsdConfigFor_String[];
  owner: string;
  unbond_time_s: number;
  utilization_method: UtilizationMethod;
  utoken: string;
  whitelist: string[];
}
export interface FeeConfigFor_String {
  immediate_withdraw_fee: Decimal;
  protocol_fee_contract: string;
  protocol_performance_fee: Decimal;
  protocol_withdraw_fee: Decimal;
}
export interface LsdConfigFor_String {
  disabled: boolean;
  lsd_type: LsdTypeFor_String;
  name: string;
}